# Reduce memory usage in CG