- Pre-allocates the vectors, improvement shown below.
- Removes the dependency of CG on KrylovSubspace (makes little sense for a 3-term recurrence method IMHO). Also: the KrylovSubspace methods are the main reason for inefficient memory usage. For the moment it was easier to fix just CG.
- Fixes a bug where the CG method starts while the initial guess is the actual solution (I still have to add a test for this)
- Adds a simple benchmark for the CG method
- Relies always on `A_mul_B!` for matrix-vector products. Since we're switching to LinearMaps this shouldn't be a problem.
- Also: removes the check whether all elements of `b` are zero. I would say CG should not be responsible for this.

Benchmark results:

```
LinearSystemsBench.cg()
```

*This PR*
```
BenchmarkTools.Trial:
  memory estimate:  45.80 MiB
  allocs estimate:  437
  --------------
  minimum time:     2.561 s (0.16% GC)
  median time:      2.580 s (0.27% GC)
  mean time:        2.580 s (0.27% GC)
  maximum time:     2.598 s (0.38% GC)
  --------------
  samples:          2
  evals/sample:     1
```

*Current implementation*
```
  BenchmarkTools.Trial:
  memory estimate:  1.53 GiB
  allocs estimate:  1041
  --------------
  minimum time:     3.173 s (9.28% GC)
  median time:      3.246 s (11.98% GC)
  mean time:        3.246 s (11.98% GC)
  maximum time:     3.320 s (14.57% GC)
  --------------
  samples:          2
  evals/sample:     1
```