@haampie 2017-06-15
Hmm, I missed the follow up of @shivin9 in #128, but still I think it is better to not use the KrylovSubspace functions and shoot just for `A_mul_B!`. CG does Krylov subspace things implicitly, and it feels hacky to initialize a KrylovSubspace of order 1.