@haampie 2017-06-15
I think the NaN result is a bug in the current implementation because the initial residual is not checked. With `b = zeros(n)` and `x0 = zeros(n)` you start out with a solution, so the iterations should not even start. The current impl. only inspect the residual after the first iteration, this PR does it before the first iteration.

If `b = zeros(n)` while `x0 = rand(n)`, then CG works just fine.