@lopezm94 2017-06-15
Removing `KrylovSubspace` should probably be in a new issue and left intact in this PR. Some solvers in this package already accept `KrylovSubspace` as input, the intention with making the `cg` implementation dependant was to enable it as input in the future. So if we are going to remove `KrylovSubspace`, it should be generalized in the whole package.

If I remember correctly there was an idea to remove `KrylovSubspace` from the package, right @jiahao? Something along the lines of using it only as a store of vectors and providing normalization functions. I could open a new issue about this. This could also improve readability for people who are more accustomed to see the matrix operations inside the solvers.