@haampie 2017-06-23
This PR doesn't remove the `KrylovSubspace` type, it removes its usage in `cg.jl`, so this PR can be merged I guess.