@lopezm94 2017-06-23
Yeah the issue is about removing it's usage in cg, not about removing the type from the package. As I explained before:

'''
Some solvers in this package already accept KrylovSubspace as input, the intention with making the cg implementation dependant was to enable it as input in the future. So if we are going to remove KrylovSubspace, it should be generalized in the whole package.
'''

So the intention was to make cg accept krylovsubspace as input in the future. If we remove it then it can't be done and it should be in a separate issue.