@barche 2017-06-25
Just for reference, I did a comparison with Trilinos.jl, resulting in the following for CG:
* Your code: 2.7 s
* Trilinos: 1.633 s

I attempted to reproduce your benchmark as accurately as possible here:
https://github.com/barche/Trilinos.jl/blob/master/benchmarks/belos.jl