@haampie 2017-06-25
@barche did you run Julia with multithreading enabled? In my experience the Julia version runs somewhat faster if you do `BLAS.set_num_threads(1)` because of all the BLAS level 1 operations.